Extract pathTuple from common
