+++ src/types/path/common.ts
@@ -1,5 +1,6 @@
 import { IsAny, IsNever } from '../utils';
 
+import { AsPathTuple, PathTuple } from './internal/pathTuple';
 import {
   AccessPattern,
   ArrayKey,
@@ -10,99 +11,6 @@ import {
   Traversable,
   UnionToIntersection,
 } from './internal/utils';
-
-/**
- * Type alias to `string` which describes a lodash-like path through an object.
- * E.g. `'foo.bar.0.baz'`
- */
-export type PathString = string;
-
-/**
- * Type which describes a path through an object
- * as a list of individual {@link Key}s.
- */
-export type PathTuple = Key[];
-
-/**
- * Type to assert that a type is a {@link PathTuple}.
- * @typeParam T - type which may be a {@link PathTuple}
- */
-export type AsPathTuple<T> = Extract<T, PathTuple>;
-
-/**
- * Type which appends a {@link Key} to the {@link PathTuple} only if it is not
- * blank, i.e. not the empty string.
- * @typeParam PT - path
- * @typeParam K  - key
- * @example
- * ```
- * AppendNonBlankKey<['foo'], 'bar'> = ['foo', 'bar']
- * AppendNonBlankKey<['foo'], ''> = ['foo']
- * ```
- */
-type AppendNonBlankKey<PT extends PathTuple, K extends Key> = K extends ''
-  ? PT
-  : [...PT, K];
-
-/**
- * Type to implement {@link SplitPathString} tail recursively.
- * @typeParam PS - remaining {@link PathString} which should be split into its
- *                 individual {@link Key}s
- * @typeParam PT - accumulator of the {@link Key}s which have been split from
- *                 the original {@link PathString} already
- */
-type SplitPathStringImpl<
-  PS extends PathString,
-  PT extends PathTuple,
-> = PS extends `${infer K}.${infer R}`
-  ? SplitPathStringImpl<R, AppendNonBlankKey<PT, K>>
-  : AppendNonBlankKey<PT, PS>;
-
-/**
- * Type to split a {@link PathString} into a {@link PathTuple}.
- * The individual {@link Key}s may be empty strings.
- * @typeParam PS  - {@link PathString} which should be split into its
- *                  individual {@link Key}s
- * @example
- * ```
- * SplitPathString<'foo'> = ['foo']
- * SplitPathString<'foo.bar.0.baz'> = ['foo', 'bar', '0', 'baz']
- * SplitPathString<'.'> = []
- * ```
- */
-export type SplitPathString<PS extends PathString> = SplitPathStringImpl<
-  PS,
-  []
->;
-
-/**
- * Type to implement {@link JoinPathTuple} tail-recursively.
- * @typeParam PT - remaining {@link Key}s which needs to be joined
- * @typeParam PS - accumulator of the already joined {@link Key}s
- */
-type JoinPathTupleImpl<
-  PT extends PathTuple,
-  PS extends PathString,
-> = PT extends [infer K, ...infer R]
-  ? JoinPathTupleImpl<AsPathTuple<R>, `${PS}.${AsKey<K>}`>
-  : PS;
-
-/**
- * Type to join a {@link PathTuple} to a {@link PathString}.
- * @typeParam PT - {@link PathTuple} which should be joined.
- * @example
- * ```
- * JoinPathTuple<['foo']> = 'foo'
- * JoinPathTuple<['foo', 'bar', '0', 'baz']> = 'foo.bar.0.baz'
- * JoinPathTuple<[]> = never
- * ```
- */
-export type JoinPathTuple<PT extends PathTuple> = PT extends [
-  infer K,
-  ...infer R
-]
-  ? JoinPathTupleImpl<AsPathTuple<R>, AsKey<K>>
-  : never;
 
 /**
  * Type which converts all keys of an object to {@link Key}s.
