+++ src/types/path/pathString.ts
@@ -1,0 +1,5 @@
+/**
+ * Type alias to `string` which describes a lodash-like path through an object.
+ * E.g. `'foo.bar.0.baz'`
+ */
+export type PathString = string;
